ApiAuth not need to have that link constant
